# Change IHubProtocol interface to support partial parsing