--- a/src/Microsoft.AspNetCore.SignalR.Client.Core/HubConnection.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Client.Core/HubConnection.cs
@@ -613,6 +613,7 @@ namespace Microsoft.AspNetCore.SignalR.Client
                     }
                     finally
                     {
+                        // The buffer was sliced up to where it was consumed, so we can just advance to the start
                         connectionState.Connection.Transport.Input.AdvanceTo(buffer.Start);
                     }
                 }