--- a/src/Microsoft.AspNetCore.SignalR.Client.Core/HubConnection.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Client.Core/HubConnection.cs
@@ -541,8 +541,10 @@ namespace Microsoft.AspNetCore.SignalR.Client
                     }
                     finally
                     {
-                        // The buffer was sliced up to where it was consumed, so we can just advance to the start
-                        _connectionState.Connection.Transport.Input.AdvanceTo(buffer.Start);
+                        // The buffer was sliced up to where it was consumed, so we can just advance to the start.
+                        // We mark examined as buffer.End so that if we didn't receive a full frame, we'll wait for more data
+                        // before yielding the read again.
+                        _connectionState.Connection.Transport.Input.AdvanceTo(buffer.Start, buffer.End);
                     }
                 }
             }
@@ -614,8 +616,10 @@ namespace Microsoft.AspNetCore.SignalR.Client
                     }
                     finally
                     {
-                        // The buffer was sliced up to where it was consumed, so we can just advance to the start
-                        connectionState.Connection.Transport.Input.AdvanceTo(buffer.Start);
+                        // The buffer was sliced up to where it was consumed, so we can just advance to the start.
+                        // We mark examined as buffer.End so that if we didn't receive a full frame, we'll wait for more data
+                        // before yielding the read again.
+                        connectionState.Connection.Transport.Input.AdvanceTo(buffer.Start, buffer.End);
                     }
                 }
             }