--- a/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/HandshakeProtocol.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/HandshakeProtocol.cs
@@ -56,7 +56,7 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
             return new JsonTextWriter(new StreamWriter(output, _utf8NoBom, 1024, leaveOpen: true));
         }
 
-        public static HandshakeResponseMessage ParseResponseMessage(ReadOnlyMemory<byte> payload)
+        public static HandshakeResponseMessage ParseResponseMessage(in ReadOnlySequence<byte> payload)
         {
             var textReader = Utf8BufferTextReader.Get(payload);
 
@@ -85,17 +85,12 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
             }
         }
 
-        public static bool TryParseRequestMessage(ReadOnlySequence<byte> buffer, out HandshakeRequestMessage requestMessage, out SequencePosition consumed, out SequencePosition examined)
+        public static bool TryParseRequestMessage(ref ReadOnlySequence<byte> buffer, out HandshakeRequestMessage requestMessage)
         {
-            if (!TryReadMessageIntoSingleMemory(buffer, out consumed, out examined, out var memory))
+            if (!TextMessageParser.TryParseMessage(ref buffer, out var payload))
             {
                 requestMessage = null;
                 return false;
-            }
-
-            if (!TextMessageParser.TryParseMessage(ref memory, out var payload))
-            {
-                throw new InvalidDataException("Unable to parse payload as a handshake request message.");
             }
 
             var textReader = Utf8BufferTextReader.Get(payload);
@@ -115,24 +110,6 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
                 Utf8BufferTextReader.Return(textReader);
             }
 
-            return true;
-        }
-
-        internal static bool TryReadMessageIntoSingleMemory(ReadOnlySequence<byte> buffer, out SequencePosition consumed, out SequencePosition examined, out ReadOnlyMemory<byte> memory)
-        {
-            var separator = buffer.PositionOf(TextMessageFormatter.RecordSeparator);
-            if (separator == null)
-            {
-                // Haven't seen the entire message so bail
-                consumed = buffer.Start;
-                examined = buffer.End;
-                memory = null;
-                return false;
-            }
-
-            consumed = buffer.GetPosition(1, separator.Value);
-            examined = consumed;
-            memory = buffer.IsSingleSegment ? buffer.First : buffer.ToArray();
             return true;
         }
     }