- These are the finishing touches before we disable batching on the
C# client and on the server. We're changing the IHubProtocol interface to
modify the input buffer with what was consumed. We're also changing it
to parse a single message at a time to be match what output writing does.
- Removes ToArray() in the optimistic case where things fit into a single buffer (for the most part)

TODO: 
- ~Run benchmarks to see if this API change helps or hurts. There are alternate patterns we can try if this hurts.~
- Add some tests for multiple buffers (we have nice helper for this in Common that I can import). This isn't really important until we disable buffering.

## Baseline:

```
             Method |          Input | HubProtocol |        Mean |      Error |     StdDev |        Op/s |  Gen 0 |  Gen 1 | Allocated |
------------------- |--------------- |------------ |------------:|-----------:|-----------:|------------:|-------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |  2,209.2 ns |  29.370 ns |  26.035 ns |   452,660.9 | 0.0343 |      - |    1120 B |
  ReadSingleMessage |   FewArguments |     MsgPack |  1,483.3 ns |  23.293 ns |  21.788 ns |   674,178.7 | 0.0248 |      - |     832 B |
  ReadSingleMessage | LargeArguments |        Json | 84,852.6 ns | 350.418 ns | 273.583 ns |    11,785.1 | 1.8311 |      - |   59016 B |
  ReadSingleMessage | LargeArguments |     MsgPack | 11,470.6 ns |  98.080 ns |  86.946 ns |    87,179.1 | 1.4038 | 0.0610 |   41688 B |
  ReadSingleMessage |  ManyArguments |        Json |  3,876.9 ns |  69.640 ns |  58.153 ns |   257,935.1 | 0.0534 |      - |    1680 B |
  ReadSingleMessage |  ManyArguments |     MsgPack |  2,983.5 ns |  54.243 ns |  50.739 ns |   335,179.1 | 0.0381 |      - |    1296 B |
  ReadSingleMessage |    NoArguments |        Json |  1,130.0 ns |  10.981 ns |   9.734 ns |   884,923.8 | 0.0229 |      - |     744 B |
  ReadSingleMessage |    NoArguments |     MsgPack |    471.7 ns |   5.856 ns |   4.890 ns | 2,119,972.0 | 0.0134 |      - |     440 B |
```


## PR

```
             Method |          Input | HubProtocol |        Mean |        Error |       StdDev |        Op/s |  Gen 0 |  Gen 1 | Allocated |
------------------- |--------------- |------------ |------------:|-------------:|-------------:|------------:|-------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |  2,292.1 ns |    37.037 ns |    34.644 ns |   436,285.5 | 0.0305 |      - |    1024 B |
  ReadSingleMessage |   FewArguments |     MsgPack |  1,536.4 ns |    10.784 ns |     9.560 ns |   650,869.1 | 0.0229 |      - |     736 B |
  ReadSingleMessage | LargeArguments |        Json | 85,141.1 ns | 1,078.927 ns |   900.953 ns |    11,745.2 | 1.9531 |      - |   58920 B |
  ReadSingleMessage | LargeArguments |     MsgPack | 11,789.1 ns |   157.994 ns |   140.057 ns |    84,824.2 | 1.4038 | 0.0610 |   41592 B |
  ReadSingleMessage |  ManyArguments |        Json |  4,016.4 ns |    78.962 ns |    69.998 ns |   248,976.4 | 0.0458 |      - |    1584 B |
  ReadSingleMessage |  ManyArguments |     MsgPack |  3,154.0 ns |    33.661 ns |    29.840 ns |   317,054.2 | 0.0381 |      - |    1200 B |
  ReadSingleMessage |    NoArguments |        Json |  1,184.8 ns |    23.713 ns |    21.021 ns |   844,039.7 | 0.0191 |      - |     648 B |
  ReadSingleMessage |    NoArguments |     MsgPack |    567.2 ns |     6.209 ns |     5.808 ns | 1,763,059.7 | 0.0105 |      - |     344 B |
 ```

It's slower and allocates less 😄. 

A few things to consider:
- The microbenchmark is testing a single buffer and there's more overhead in a `ReadOnlyMemory<byte>` vs a `ReadOnlySequence<byte>`. Slicing a ReadOnlySequence is slower and that could explain the overhead.
- Since we're parsing a single message, we can avoid allocating a list (the list could be cached per connection and constantly cleared after parsing).

Contributes to #1508