@BrennanConroy 2018-03-28
Before
```
             Method |          Input | HubProtocol |         Mean |        Error |       StdDev |        Op/s |  Gen 0 | Allocated |
------------------- |--------------- |------------ |-------------:|-------------:|-------------:|------------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |   2,676.2 ns |    57.837 ns |    71.029 ns |   373,665.3 | 0.0076 |    1120 B |
 WriteSingleMessage |   FewArguments |        Json |   2,990.4 ns |    73.083 ns |   212.028 ns |   334,402.1 | 0.0648 |    6608 B |
  ReadSingleMessage |   FewArguments |     MsgPack |   1,496.2 ns |    26.538 ns |    24.823 ns |   668,378.8 | 0.0076 |     832 B |
 WriteSingleMessage |   FewArguments |     MsgPack |   1,513.8 ns |    35.312 ns |    45.915 ns |   660,609.0 | 0.0134 |    1224 B |
  ReadSingleMessage | LargeArguments |        Json | 105,269.2 ns | 2,046.561 ns | 2,935.116 ns |     9,499.5 | 0.4883 |   59016 B |
 WriteSingleMessage | LargeArguments |        Json |  73,056.8 ns | 1,433.446 ns | 1,650.758 ns |    13,688.0 | 0.8545 |   94952 B |
  ReadSingleMessage | LargeArguments |     MsgPack |  13,042.7 ns |   259.305 ns |   574.602 ns |    76,671.5 | 0.4272 |   41688 B |
 WriteSingleMessage | LargeArguments |     MsgPack |  34,038.5 ns |   900.969 ns | 2,656.528 ns |    29,378.5 | 1.1597 |  113944 B |
  ReadSingleMessage |  ManyArguments |        Json |   4,189.6 ns |    54.910 ns |    45.852 ns |   238,683.7 | 0.0153 |    1680 B |
 WriteSingleMessage |  ManyArguments |        Json |   4,763.7 ns |    95.121 ns |   240.382 ns |   209,921.9 | 0.0839 |    7864 B |
  ReadSingleMessage |  ManyArguments |     MsgPack |   2,732.0 ns |    53.246 ns |    71.082 ns |   366,029.5 | 0.0114 |    1296 B |
 WriteSingleMessage |  ManyArguments |     MsgPack |   2,764.3 ns |    53.836 ns |    86.935 ns |   361,752.2 | 0.0114 |    1496 B |
  ReadSingleMessage |    NoArguments |        Json |   1,151.6 ns |    22.893 ns |    36.968 ns |   868,386.3 | 0.0076 |     744 B |
 WriteSingleMessage |    NoArguments |        Json |   1,795.6 ns |    58.759 ns |   172.329 ns |   556,914.5 | 0.0648 |    6176 B |
  ReadSingleMessage |    NoArguments |     MsgPack |     403.2 ns |     7.942 ns |     9.753 ns | 2,480,351.1 | 0.0048 |     440 B |
 WriteSingleMessage |    NoArguments |     MsgPack |     647.7 ns |    12.991 ns |    29.322 ns | 1,543,914.9 | 0.0095 |     976 B |
```

After
```
	 
	         Method |          Input | HubProtocol |         Mean |       Error |      StdDev |        Op/s |  Gen 0 |  Gen 1 | Allocated |
------------------- |--------------- |------------ |-------------:|------------:|------------:|------------:|-------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |   2,769.2 ns |    55.71 ns |    74.37 ns |   361,111.8 | 0.0076 |      - |    1024 B |
 WriteSingleMessage |   FewArguments |        Json |   2,857.5 ns |    56.91 ns |   160.51 ns |   349,960.0 | 0.0687 |      - |    6608 B |
  ReadSingleMessage |   FewArguments |     MsgPack |   1,562.9 ns |    31.14 ns |    47.56 ns |   639,842.4 | 0.0076 |      - |     736 B |
 WriteSingleMessage |   FewArguments |     MsgPack |   1,456.1 ns |    30.03 ns |    54.15 ns |   686,783.2 | 0.0134 |      - |    1224 B |
  ReadSingleMessage | LargeArguments |        Json | 105,728.8 ns | 2,087.19 ns | 3,370.42 ns |     9,458.2 | 0.4883 |      - |   58920 B |
 WriteSingleMessage | LargeArguments |        Json |  73,448.6 ns | 1,454.68 ns | 2,349.04 ns |    13,615.0 | 0.8545 |      - |   94952 B |
  ReadSingleMessage | LargeArguments |     MsgPack |  13,256.6 ns |   262.96 ns |   692.73 ns |    75,434.1 | 0.4425 |      - |   41592 B |
 WriteSingleMessage | LargeArguments |     MsgPack |  34,021.8 ns |   811.35 ns | 2,392.28 ns |    29,392.9 | 1.2207 | 0.0610 |  113944 B |
  ReadSingleMessage |  ManyArguments |        Json |   4,361.4 ns |    85.34 ns |    94.86 ns |   229,285.4 | 0.0153 |      - |    1584 B |
 WriteSingleMessage |  ManyArguments |        Json |   4,747.4 ns |    94.88 ns |   258.12 ns |   210,640.9 | 0.0839 |      - |    7864 B |
  ReadSingleMessage |  ManyArguments |     MsgPack |   3,115.1 ns |    62.05 ns |   100.20 ns |   321,013.5 | 0.0114 |      - |    1200 B |
 WriteSingleMessage |  ManyArguments |     MsgPack |   2,550.1 ns |    48.91 ns |    54.36 ns |   392,137.1 | 0.0153 |      - |    1496 B |
  ReadSingleMessage |    NoArguments |        Json |   1,181.2 ns |    13.90 ns |    11.60 ns |   846,575.1 | 0.0057 |      - |     648 B |
 WriteSingleMessage |    NoArguments |        Json |   1,770.6 ns |    38.19 ns |   112.01 ns |   564,770.7 | 0.0648 |      - |    6176 B |
  ReadSingleMessage |    NoArguments |     MsgPack |     517.8 ns |    10.33 ns |    26.67 ns | 1,931,337.4 | 0.0038 |      - |     344 B |
 WriteSingleMessage |    NoArguments |     MsgPack |     661.3 ns |    13.12 ns |    31.44 ns | 1,512,119.3 | 0.0105 |      - |     976 B |
```

Reads got a little slower