@davidfowl 2018-03-28
I think we're going to have to live with this performance hit. The microbenchmark highlights the difference between ROMemory and ROSequence